requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0
pandas>=2.0.0
